llama/requirements.txt
# Core dependencies
torch>=2.1.0
transformers>=4.38.0
accelerate>=0.26.0
datasets>=2.16.0
peft>=0.8.0
bitsandbytes>=0.41.0
trl>=0.7.10

# Evaluation and utilities
evaluate>=0.4.1
scikit-learn>=1.3.0
nltk>=3.8.1
textstat>=0.7.3

# Configuration and logging
pyyaml>=6.0
wandb>=0.16.0
tensorboard>=2.15.0

# Optional but recommended
flash-attn>=2.5.0
deepspeed>=0.12.0